1.source-to-pay-engine/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
requests
google-generativeai
pydantic
pydantic-settings
python-dotenv
tenacity
